refactor(ConnectionGeneator): Connection 객체를 생성하는 역할을 외부 클래스로 분리
