+++ docs/README.md
@@ -11,6 +11,9 @@
 - 체스 게임은 현재 남아 있는 말에 대한 점수를 구할 수 있어야 한다.
 - "status" 명령을 입력하면 각 진영의 점수를 출력하고 어느 진영이 이겼는지 결과를 볼 수 있어야 한다
 
+- 애플리케이션을 재시작하더라도 이전에 하던 체스 게임을 다시 시작할 수 있어야 한다.
+- DB를 적용할 때 도메인 객체의 변경을 최소화해야한다.
+
 # 기능 구현 목록
 
 ## 게임 명령을 입력하는 기능
