+++ src/main/java/repository/PieceRepository.java
@@ -2,38 +2,20 @@ package repository;
 
 import static repository.ColorMapper.getColorByFieldName;
 import static repository.PieceRoleMapper.getPieceRoleByFieldName;
-import static repository.PropertiesGenerator.properties;
 
 import domain.piece.Color;
 import domain.piece.Piece;
 import domain.piece.piecerole.PieceRole;
 import java.sql.Connection;
-import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
 public class PieceRepository {
-    private static final String SERVER = properties().getProperty("server");
-    private static final String DATABASE = properties().getProperty("database");
-    private static final String OPTION = properties().getProperty("option");
-    private static final String USERNAME = properties().getProperty("username");
-    private static final String PASSWORD = properties().getProperty("password");
-
-    public Connection getConnection() {
-        try {
-            return DriverManager.getConnection("jdbc:mysql://" + SERVER + "/" + DATABASE + OPTION, USERNAME, PASSWORD);
-        } catch (final SQLException exception) {
-            System.err.println("DB 연결 오류:" + exception.getMessage());
-            exception.printStackTrace();
-            return null;
-        }
-    }
-
     public int save(final Piece piece) {
         String query = "INSERT INTO piece (piece_role, color) VALUES (?, ?)";
 
-        Connection connection = getConnection();
+        Connection connection = ConnectionGenerator.getConnection();
         try {
             PreparedStatement preparedStatement = connection.prepareStatement(query);
             preparedStatement.setString(1, piece.pieceRoleName());
@@ -48,7 +30,7 @@ public class PieceRepository {
     public Piece findByPieceId(final String pieceId) {
         String query = "SELECT * FROM piece WHERE piece_id = ?";
 
-        Connection connection = getConnection();
+        Connection connection = ConnectionGenerator.getConnection();
         try {
             PreparedStatement preparedStatement = connection.prepareStatement(query);
             preparedStatement.setString(1, pieceId);
